finish project

+++ src/main/java/com/todolist/todolist/task/TaskController.java
@@ -1,10 +1,13 @@
 package com.todolist.todolist.task;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,6 +16,8 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.todolist.todolist.utils.Utils;
 
 import jakarta.servlet.http.HttpServletRequest;
 
@@ -24,12 +29,22 @@ public class TaskController {
   private ITaskRepository taskRepository;
 
   @PostMapping("/create")
-  public TaskModel create(@RequestBody TaskModel taskModel, HttpServletRequest request) {
+  public ResponseEntity<Object> create(@RequestBody TaskModel taskModel, HttpServletRequest request) {
     var idUser = (UUID) request.getAttribute("idUser");
     taskModel.setIdUser(idUser);
+    var currentDate = LocalDateTime.now();
+    var startTime = taskModel.getStartAt();
     
+    if(currentDate.isAfter(startTime)) {
+      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Start time must be before current time");
+    }
+
+    if(startTime.isBefore(taskModel.getEndAt())) {
+      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("End time must be after start time");
+    }
+
     var task = this.taskRepository.save(taskModel);
-    return task;
+    return ResponseEntity.status(HttpStatus.CREATED).body(task);
   }
 
   @GetMapping("")
@@ -38,8 +53,10 @@ public class TaskController {
     return tasks;
   }
 
-  @GetMapping("user/{idUser}")
-  public List<TaskModel> fetchUsersTasks(@PathVariable UUID idUser) {
+  @GetMapping("user/")
+  public List<TaskModel> fetchUsersTasks(HttpServletRequest resquest) {
+    var idUser = (UUID) resquest.getAttribute("idUser");
+
     var tasks = this.taskRepository.findByIdUser(idUser);
     return tasks;
   }
@@ -51,20 +68,36 @@ public class TaskController {
   }
 
   @PutMapping("{id}")
-  public TaskModel updateTask(@PathVariable UUID id, @RequestBody TaskModel taskModel) {
-    var task = this.taskRepository.findById(id).get();
+  public ResponseEntity<Object> updateTask(@PathVariable UUID id, @RequestBody TaskModel taskModel,HttpServletRequest resquest) {
+    var task = this.taskRepository.findById(id).orElse(null);
+    var idUser = (UUID) resquest.getAttribute("idUser");
 
-    task.setTitle(taskModel.getTitle());
-    task.setDescription(taskModel.getDescription());
-    task.setPriority(taskModel.getPriority());
-    task.setStartAt(taskModel.getStartAt());
-    task.setEndAt(taskModel.getEndAt());
+    if(task == null) {
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Task not found");
+    }
 
-    return this.taskRepository.save(task);
+    if(!task.getIdUser().equals(idUser)) {
+      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("This task does not belong to you");
+    }
+    Utils.copyNonNullProperties(taskModel, task);
+
+   var updatedTask= this.taskRepository.save(taskModel);
+   return ResponseEntity.status(HttpStatus.OK).body(updatedTask);
   }
 
   @DeleteMapping("{id}")
-  public void delete(@PathVariable UUID id) {
+  public ResponseEntity<Object> delete(@PathVariable UUID id, HttpServletRequest resquest) {
+    var task = this.taskRepository.findById(id).get();
+    var idUser = (UUID) resquest.getAttribute("idUser");
+
+    if(task == null) {
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Task not found");
+    }
+      
+    if(!task.getIdUser().equals(idUser)) {
+      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("This task does not belong to you");
+    }
     this.taskRepository.deleteById(id);
+    return ResponseEntity.status(HttpStatus.OK).body("Task deleted");
   }
 }
